version H2

+++ src/main/java/com/gestion/libreria/repository/libreriaRepository.java
@@ -3,5 +3,5 @@ package com.gestion.libreria.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.gestion.libreria.entity.libro;
 
-public interface libreriaRepository extends JpaRepository<libro,Long> {
+public interface libreriaRepository extends JpaRepository<libro,Integer> {
 }
